# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: LeptonPad API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /content:
    get:
      operationId: listContent
      tags: [content]
      summary: List content feed
      parameters:
        - name: categories
          in: query
          schema:
            type: string
          description: Comma-separated category slugs to filter by
        - name: type
          in: query
          schema:
            type: string
            enum: [article, audio, video]
          description: Filter by content type
        - name: creatorId
          in: query
          schema:
            type: string
          description: Filter by creator Clerk user ID
        - name: q
          in: query
          schema:
            type: string
          description: Search titles, previews, and creator names
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: Paginated content list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContentListResponse"
    post:
      operationId: createContent
      tags: [content]
      summary: Create new content
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContentInput"
      responses:
        "201":
          description: Created content
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Content"
        "400":
          description: Invalid input
        "401":
          description: Unauthorized
  /content/{id}:
    get:
      operationId: getContent
      tags: [content]
      summary: Get single content item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Content item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContentDetail"
        "404":
          description: Not found
    put:
      operationId: updateContent
      tags: [content]
      summary: Update content
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContentUpdate"
      responses:
        "200":
          description: Updated content
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Content"
        "401":
          description: Unauthorized
        "404":
          description: Not found
    delete:
      operationId: deleteContent
      tags: [content]
      summary: Delete content
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "401":
          description: Unauthorized
        "404":
          description: Not found
  /content/{id}/next:
    get:
      operationId: getNextContent
      tags: [content]
      summary: Get next content item in the same category
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Next content item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Content"
        "204":
          description: No next item found
  /content/{id}/view:
    post:
      operationId: trackContentView
      tags: [content]
      summary: Track a view on content
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: View tracked
...
  /earnings/content:
    get:
      operationId: getEarningsByContent
      tags: [earnings]
      summary: Get earnings breakdown per content item
      responses:
        "200":
          description: Per-content earnings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ContentEarnings"
        "401":
          description: Unauthorized
...
components:
  schemas:
...
    Content:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        type:
          type: string
          enum: [article, audio, video]
        categorySlug:
          type: string
        categoryName:
          type: string
        price:
          type: number
        creatorId:
          type: string
        creatorName:
          type: string
        creatorImageUrl:
          type: ["string", "null"]
        creatorVerified:
          type: boolean
        previewText:
          type: ["string", "null"]
        coverImageUrl:
          type: ["string", "null"]
        audioUrl:
          type: ["string", "null"]
        videoUrl:
          type: ["string", "null"]
        viewCount:
          type: integer
        purchaseCount:
          type: integer
        published:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - title
        - type
        - categorySlug
        - categoryName
        - price
        - creatorId
        - creatorName
        - creatorImageUrl
        - creatorVerified
        - previewText
        - coverImageUrl
        - audioUrl
        - videoUrl
        - viewCount
        - purchaseCount
        - published
        - createdAt
    ContentDetail:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        type:
          type: string
          enum: [article, audio, video]
        categorySlug:
          type: string
        categoryName:
          type: string
        price:
          type: number
        creatorId:
          type: string
        creatorName:
          type: string
        creatorImageUrl:
          type: ["string", "null"]
        body:
          type: ["string", "null"]
        previewText:
          type: ["string", "null"]
        coverImageUrl:
          type: ["string", "null"]
        audioUrl:
          type: ["string", "null"]
        videoUrl:
          type: ["string", "null"]
        viewCount:
          type: integer
        purchaseCount:
          type: integer
        readingTimeMinutes:
          type: ["integer", "null"]
        published:
          type: boolean
        createdAt:
          type: string
        hasAccess:
          type: boolean
        creatorVerified:
          type: boolean
      required:
        - id
        - title
        - type
        - categorySlug
        - categoryName
        - price
        - creatorId
        - creatorName
        - creatorImageUrl
        - creatorVerified
        - body
        - previewText
        - coverImageUrl
        - audioUrl
        - videoUrl
        - viewCount
        - purchaseCount
        - readingTimeMinutes
        - published
        - createdAt
        - hasAccess
    ContentInput:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        type:
          type: string
          enum: [article, audio, video]
        categorySlug:
          type: string
        body:
          type: string
        previewText:
          type: string
        coverImageUrl:
          type: string
          description: Uploaded file path or absolute URL
        audioUrl:
          type: string
        videoUrl:
          type: string
        price:
          type: number
          minimum: 0
        published:
          type: boolean
      required:
        - title
        - type
        - categorySlug
        - price
        - coverImageUrl
    ContentUpdate:
      type: object
      properties:
        title:
          type: string
        body:
          type: string
        previewText:
          type: string
        coverImageUrl:
          type: string
          description: Uploaded file path or absolute URL
        audioUrl:
          type: string
        videoUrl:
          type: string
        price:
          type: number
          minimum: 0
        published:
          type: boolean
    ContentListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Content"
        total:
          type: integer
        offset:
          type: integer
        limit:
          type: integer
      required:
        - items
        - total
        - offset
        - limit
...
    ContentEarnings:
      type: object
      properties:
        contentId:
          type: integer
        title:
          type: string
        type:
          type: string
        price:
          type: number
        views:
          type: integer
        purchases:
          type: integer
        conversionRate:
          type: number
        totalEarned:
          type: number
        hasSuggestion:
          type: boolean
        publishedAt:
          type: string
      required:
        - contentId
        - title
        - type
        - price
        - views
        - purchases
        - conversionRate
        - totalEarned
        - hasSuggestion
        - publishedAt
...